+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Adhered to stricter abstract base classes in transformer subpackage
+
 ## [1.0.1] - 2026-04-16
 - Removed redundant parameter from cross-attention feature mixer