+++ CHANGELOG.md
@@ -4,17 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-
+## [2026-04-20]
 ### Fixed
 - Adhered to stricter abstract base classes in transformer subpackage
 - Norm-first logic in EncoderLayer
+
+### Changed
+- Factored merge-mask logic out of encoder forward
+
 
 ## [1.0.1] - 2026-04-16
 - Removed redundant parameter from cross-attention feature mixer
 
 ## [1.0.0] - 2026-04-15
-
 ### Changed
 - New AWS and GCP cloud APIs
 - funcflow Map now has a "flat" option